refactor: move edit_op + edit_op_apply into homeboy-engine-primitives (#8425)
Continue the floor migration. Both are pure edit-application primitives:
- edit_op (EditOp / InsertAnchor vocabulary): zero crate:: deps.
- edit_op_apply (+ its imports submodule): only Language (now a primitive) and
  homeboy_error::Result.

After the #8431 split, edit_op_apply is fully decoupled from the fixer/refactor
tagging layer (it takes plain &[EditOp]), so nothing about it is core-specific
anymore. Move the module + edit_op_apply/imports.rs into
homeboy-engine-primitives; repoint super::edit_op -> crate::edit_op,
crate::engine::language -> crate::language, crate::error -> homeboy_error. Bridge
both back through engine/mod.rs's existing re-export so the ~6 consumers keep
resolving crate::engine::{edit_op, edit_op_apply} unchanged.

Depends on the Language migration (edit_op_apply uses Language); stacked on that
change.

Verified: primitives + core compile clean; 45 edit_op tests pass;
`cargo check --workspace --tests` green.

Refs #8425
