+++ crates/homeboy-core/src/engine/mod.rs
@@ -6,16 +6,14 @@
 // crate. Re-exported here so existing `crate::engine::{shell,command,...}`
 // call sites keep working unchanged.
 pub use homeboy_engine_primitives::{
-    baseline, codebase_scan, command, detail_output, identifier, language, output_parse, shell,
-    template, text, validation,
+    baseline, codebase_scan, command, detail_output, edit_op, edit_op_apply, identifier, language,
+    output_parse, shell, template, text, validation,
 };
 // local_files was `pub(crate)` in-tree; preserve that visibility across the
 // crate boundary rather than widening it via the `pub use` above.
 pub(crate) use homeboy_engine_primitives::local_files;
 
 pub mod cli_tool;
-pub mod edit_op;
-pub mod edit_op_apply;
 pub mod execution_context;
 pub mod executor;
 pub mod format_write;
